Move 4byte's phf generation out of build.rs
