+++ etk-4byte/Cargo.toml
@@ -13,17 +13,19 @@ categories = ["cryptography::cryptocurrencies", "development-tools"]
 
 
 [features]
-generate = ["reqwest", "tokio", "serde"]
+generate = ["reqwest", "tokio", "serde", "phf_codegen"]
 
 [dependencies]
 reqwest = { version = "0.11.2", optional = true, features = [ "json" ]}
 tokio = { version = "1.3.0", optional = true, features = [ "macros", "sync" ] }
 serde = { version = "1.0", optional = true, features = [ "derive" ] }
+phf_codegen = { version = "0.9.0", optional = true }
 phf = "0.9.0"
-
-[build-dependencies]
-phf_codegen = "0.9.0"
 
 [[bin]]
 name = "etk-4byte-generate"
+required-features = [ "generate" ]
+
+[[bin]]
+name = "etk-4byte-phf"
 required-features = [ "generate" ]
